src/requirements.txt
langchain>=1.2.0
langchain-groq>=0.3.0
langchain-core>=1.2.0
langgraph>=1.1.0
python-dotenv>=1.0.0